# api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for the Kootenai cybersecurity education platform\nOrchestrates lab environments across Proxmox and CloudStack",
        "title": "Kootenai API",
        "contact": {
            "name": "Project repository",
            "url": "https://github.com/toddbartholow/kootenai"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/pathways/{pathwayID}/next-lab": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the next recommended lab to complete in a specific pathway",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recommendations"
                ],
                "summary": "Get Next Lab in Pathway",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pathway ID or slug",
                        "name": "pathwayID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_toddbartholow_kootenai_api_internal_recommendation.LabRecommendation"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/internal_server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No next lab found or not enrolled",
                        "schema": {
                            "$ref": "#/definitions/internal_server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/recommendations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get personalized lab and pathway recommendations for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recommendations"
                ],
                "summary": "Get Recommendations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum number of recommendations per category (default: 5)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_toddbartholow_kootenai_api_internal_recommendation.RecommendationsResponse"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/internal_server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_toddbartholow_kootenai_api_internal_recommendation.LabRecommendation": {
            "type": "object",
            "properties": {
                "difficulty": {
                    "type": "string"
                },
                "durationMinutes": {
                    "type": "integer"
                },
                "labDescription": {
                    "type": "string"
                },
                "labName": {
                    "type": "string"
                },
                "labSlug": {
                    "type": "string"
                },
                "labTemplateId": {
                    "type": "string"
                },
                "maxPoints": {
                    "type": "integer"
                },
                "moduleId": {
                    "type": "string"
                },
                "moduleName": {
                    "type": "string"
                },
                "pathwayId": {
                    "description": "Context for pathway-based recommendations",
                    "type": "string"
                },
                "pathwayName": {
                    "type": "string"
                },
                "priority": {
                    "description": "Lower is higher priority",
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_toddbartholow_kootenai_api_internal_recommendation.RecommendationType"
                }
            }
        },
        "github_com_toddbartholow_kootenai_api_internal_recommendation.PathwayRecommendation": {
            "type": "object",
            "properties": {
                "coverImageUrl": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "difficulty": {
                    "type": "string"
                },
                "estimatedHours": {
                    "type": "integer"
                },
                "icon": {
                    "type": "string"
                },
                "labCount": {
                    "type": "integer"
                },
                "moduleCount": {
                    "type": "integer"
                },
                "pathwayId": {
                    "type": "string"
                },
                "pathwayName": {
                    "type": "string"
                },
                "pathwaySlug": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_toddbartholow_kootenai_api_internal_recommendation.RecommendationType"
                }
            }
        },
        "github_com_toddbartholow_kootenai_api_internal_recommendation.RecommendationType": {
            "type": "string",
            "enum": [
                "next_in_pathway",
                "continue_progress",
                "new_pathway",
                "similar_difficulty",
                "popular"
            ],
            "x-enum-varnames": [
                "TypeNextInPathway",
                "TypeContinueProgress",
                "TypeNewPathway",
                "TypeSimilarDifficulty",
                "TypePopular"
            ]
        },
        "github_com_toddbartholow_kootenai_api_internal_recommendation.RecommendationsResponse": {
            "type": "object",
            "properties": {
                "labs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_toddbartholow_kootenai_api_internal_recommendation.LabRecommendation"
                    }
                },
                "pathways": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_toddbartholow_kootenai_api_internal_recommendation.PathwayRecommendation"
                    }
                }
            }
        },
...
        "internal_server.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid credentials"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}